formatting_engine/requirements.txt
pymupdf
pdfplumber
docling
groq
anthropic
python-dotenv
nougat-ocr
albumentations<2
